Added an interactive console project
Interactive console is a console api that looks similar to the Curses
console, input will be read on the last line and output will be the
upper part of the console.  There will be a status bar separating input
and output.  It is class based, so you can have multiple windows
running.

TODO:
- Static methods that work on the currently active window.

Added correct licensing

Added licensing to the head of all files.

Fixed git modifications

Git modified all the files. Fixed.